Add files via upload
A blood donation Management system created using HTML,CSS,PHP and MYSQL. It gives general information for users about the do's, dont's, before and after a blood donation. Along with that, it takes the details of donors and recipients and exchanges the information to a blood bank, which will then further contact the donor/recipient. Hence this mini project was developed in the aim of Saving lives!(Uses forms for data input and MYSQL database to store the records and transfer it to a blood bank).